+++ R/ggplot2.R
@@ -22,6 +22,7 @@ scatterplot_ggplot2 <- function(x,
   
   control <- c(control, list(...))  
   control <- .get_parameters(control, list(
+    main = paste("Scatter plot for", length(x), class(x)),
     colors = default_colors(2), 
     jitter = NA,
     engine = "ggplot2"
@@ -55,11 +56,12 @@ scatterplot_ggplot2 <- function(x,
   jitter <- jitter[1]
   if(is.na(jitter) && any(duplicated(q[,measure]))) {
     message("To reduce overplotting, jitter is added! Use jitter = 0 to prevent jitter.")   
-    jitter <- .1
+    jitter <- .jitter_default
   }
   
   if(!is.na(jitter) && jitter>0) 
-    for(m in measure) q[[m]] <- jitter(q[[m]], factor = jitter)
+    for(m in measure) 
+      if(is.numeric(q[[m]])) q[[m]] <- jitter(q[[m]], factor = jitter)
 
   if(is.na(shading)) shading <- NULL
   p <- ggplot(q, aes_string(measure[1], y = measure[2], color = shading)) +
@@ -71,7 +73,8 @@ scatterplot_ggplot2 <- function(x,
     else
       p <- p + scale_color_discrete()
   }
-  p
+  
+  p + ggtitle(control$main) + theme_linedraw()
   
 }
 
@@ -80,6 +83,7 @@ matrix_ggplot2 <- function(x,
   
   control <- c(control, list(...))  
   control <- .get_parameters(control, list(
+    main = paste("Matrix for", length(x), "rules"),
     colors = default_colors(2), 
     reorder = "measure",
     max = 1000,
@@ -105,9 +109,9 @@ matrix_ggplot2 <- function(x,
   cat(measure) 
   
   
-  m <- rulesAsMatrix(x, measure = measure)
-  m_s <- rulesAsMatrix(x, "support")
-  m_c <- rulesAsMatrix(x, "confidence")
+  m <- rules2matrix(x, measure = measure)
+  m_s <- rules2matrix(x, "support")
+  m_c <- rules2matrix(x, "confidence")
   
   reorderTypes <- c("none", "measure", "support/confidence", "similarity")
   reorderType <- pmatch(control$reorder , reorderTypes, nomatch = 0)
@@ -151,14 +155,130 @@ matrix_ggplot2 <- function(x,
   dimnames(m) <- list(seq_len(nrow(m)), seq_len(ncol(m)))
    
   # NOTE: nullify variables used for non-standard evaluation for tidyverse/ggplot2 below
-  RHS <- LHS <- value <- NULL
+  RHS <- LHS  <- NULL
   
   d <- m %>% as_tibble() %>% dplyr::mutate(RHS = seq_len(nrow(m))) %>% 
-    pivot_longer(cols = -c(RHS), names_to = "LHS")  
+    pivot_longer(cols = -c(RHS), names_to = "LHS", values_to = measure)  
   d$LHS <- as.integer(d$LHS)
   
-  ggplot(d, aes_string(x = 'LHS', y = 'RHS', fill = 'value')) + geom_tile() +
-    scale_fill_gradient(low=colors[1], high=colors[2], na.value = 0) 
+  ggplot(d, aes_string(x = 'LHS', y = 'RHS', fill = measure)) + geom_raster() +
+    scale_fill_gradient(low=colors[1], high=colors[2], na.value = 0) +
+    ggtitle(control$main) + 
+    theme_linedraw()
 }
+
+
+graph_ggplot2 <- function(x, measure = "support", shading = "lift", 
+  control = NULL, ...) {
+  
+  if(is.na(shading)) shading <- NULL
+  
+  ### NULLify stuff for CRAN
+  label <- y <- xend <- yend <- NULL
+  
+  control <- c(control, list(...))  
+  control <- .get_parameters(control, list(
+    #main = paste("Graph for", length(x), "rules"),
+    layout = igraph::nicely(),
+    edges = ggnetwork::geom_edges(color = "grey80", 
+      arrow = arrow(length = unit(6, "pt"), type = "closed"), alpha = .7),
+    nodes = ggnetwork::geom_nodes(aes_string(size = measure, color = shading), na.rm = TRUE),
+    nodetext = ggnetwork::geom_nodetext(aes(label = label)),
+    colors = default_colors(2), 
+    engine = "ggplot2", 
+    max = 100
+  ))
+  
+  if(length(x) > control$max) {
+    warning("Too many rules supplied. Only plotting the best ", 
+      control$max, " rules using ", shading, 
+      " (change control parameter max if needed)", call. = FALSE)
+    x <- tail(x, n = control$max, by = shading, decreasing = FALSE)
+  }
+  
+  
+  g <- associations2igraph(x)
+  n <- ggnetwork::fortify(g, layout = control$layout)
+  
+  ggplot(n, aes(x = x, y = y, xend = xend, yend = yend)) +
+    control$edges +
+    control$nodes +
+    control$nodetext + 
+    scale_color_gradient(low = control$colors[2], high = control$colors[1], na.value = 0) +
+    ggnetwork::theme_blank()
+}
+
+grouped_matrix_ggplot2 <- function(x, 
+  measure = c("support"), shading = "lift", control = NULL, ...) {
+  
+  control <- c(control, list(...))  
+  control <- .get_parameters(control, list(
+    k = 20,
+    rhs_max = 10,
+    lhs_items = 2,
+    aggr.fun=mean, 
+    col = default_colors(2),
+    max.shading = NA,
+    reverse = TRUE,
+    engine = "ggplot2"
+  ))
+  
+  ### get the clustering
+  p <- grouped_matrix_int(x, measure, shading, control, plot = FALSE)  
+
+  ## get most important item in the lhs
+  f <- lapply(split(p$rules, p$cl), FUN = function(r) itemFrequency(lhs(r), 
+    type = "absolute"))
+  
+  ## divide by sum to find most important item...
+  f <- lapply(f, "/", itemFrequency(lhs(p$rules), type = "absolute")+1L)
+  
+  most_imp_item <- lapply(f, FUN = 
+      function(x) {
+        items <- sum(x>0)
+        if(items==0) { "" }
+        else if(control$lhs_items<1){ 
+          paste(items, "items") 
+        } else if(items>control$lhs_items){
+          paste(paste(names(x[head(order(x, decreasing = TRUE), n = control$lhs_items)]), 
+            collapse = ", "), ", +", items-control$lhs_items, " items", sep="")
+        }else{
+          paste(names(x[head(order(x, decreasing = TRUE), n = items)]), 
+            collapse = ", ")
+        }
+      })
+  
+  s <- p$sAggr
+  m <- p$mAggr
+  colnames(s) <- paste(
+    paste(format(table(p$cl)), " rules: ", '{',most_imp_item, '}', sep=''))
+  
+  if(control$reverse) p$order[[1]] <- rev(p$order[[1]])
+  s <- permute(s, p$order)
+  m <- permute(m, p$order)
+  
+  df <- data.frame(
+    LHS = rep(ordered(colnames(s), levels = colnames(s)), times = nrow(s)), 
+    RHS = rep(ordered(rownames(s), levels = rownames(s)), each = ncol(s)), 
+    support = as.vector(t(m)), 
+    measure = as.vector(t(s))
+  )
+
+  
+  ### NULLify for CRAN
+  LHS <- RHS <- NULL
+  
+  p <- ggplot(df, aes(x = LHS, y = RHS, size = support, color = measure)) +
+    geom_point(na.rm = TRUE) + 
+    scale_color_gradient(low = control$col[2], high = control$col[1]) + 
+    labs(color = p$shading) +
+    theme_linedraw() +
+    theme(axis.text.x=element_text(angle=90, hjust=0, vjust = .5)) +
+    scale_x_discrete(position = "top") 
+
+  if(control$engine == "htmlwidget") p <- plotly::ggplotly(p)
+  p
+}
+