+++ R/plotly.R
@@ -18,53 +18,11 @@
 
 # plotly interactive plots using d3.js
 
-
-plotly_arules <- function(x, method = "scatterplot", 
-  measure = c("support", "confidence"), shading = "lift", 
-  max = 1000, ...) {
-  
-  .Deprecated("plot")
-
-  .plotly_arules(x, method, measure, shading, 1000, ...)
-}
-
-.plotly_arules <- function(x, method = "scatterplot", 
-  measure = c("support", "confidence"), shading = "lift", 
-  max = 1000, ...) {
-  
-  if(!is(x, "associations")) stop("x has to be a set of rules.")
-  
-  methods <- c("scatterplot", "two-key plot", "matrix")
-  
-  if(length(x)<1) stop("x contains 0 rules!")
-  
-  if(is.null(method)) methodNr <- 1
-  else methodNr <- pmatch(tolower(method), tolower(methods))
-  if(is.na(methodNr)) stop (paste("Unknown method:",sQuote(method)))
-  
-  # two-key plot is a scatterplot with order shading
-  if(methodNr==2) {
-    shading <- "order"
-    methodNr <- 1
-  }
-  
-  #quality(x)[["order"]] <- size(x) 
-  qnames <- names(quality(x))
-  measure <- qnames[pmatch(measure, qnames, duplicates.ok = TRUE)]
-  shading <- qnames[pmatch(shading, qnames)]
-  
-  if(methodNr == 1) 
-    .plotly_scatter(x, measure, shading, max = max, ...)
-  else 
-    .plotly_matrix(x, shading, max = max, ...)
-}
-
 ## Interface used by plot
 scatterplot_plotly <- function(x,
   measure = measure, shading = shading, control = control, ...) {
   
-  control <- c(control, list(...))  
-  
+  control <- c(control, list(...))
   control <- .get_parameters(control, list(
     interactive = TRUE,
     engine = "htmlwidget",
@@ -136,7 +94,7 @@ scatterplot_plotly <- function(x,
   jitter <- jitter[1]
   if(is.na(jitter) && any(duplicated(q[,measure]))) {
     message("To reduce overplotting, jitter is added! Use jitter = 0 to prevent jitter.")   
-    jitter <- .1
+    jitter <- .jitter_default
   }
   
   if(!is.na(jitter) && jitter>0) 
@@ -169,11 +127,9 @@ scatterplot_plotly <- function(x,
     )
 }
 
-## interface for plot
 matrix_plotly <- function(x, measure, shading, control, ...) {
   
   control <- c(control, list(...))  
-  
   control <- .get_parameters(control, list(
     interactive = TRUE,
     engine = "htmlwidget",
@@ -202,10 +158,10 @@ matrix_plotly <- function(x, measure, shading, control, ...) {
     x <- tail(x, n = max, by = measure, decreasing = FALSE)
   }
   
-  m <- rulesAsMatrix(x, measure = measure, itemSep= ',<BR>&nbsp;&nbsp;', 
+  m <- rules2matrix(x, measure = measure, itemSep= ',<BR>&nbsp;&nbsp;', 
     setStart = '<B>{', setEnd = '}</B>')
-  m_s <- rulesAsMatrix(x, "support")
-  m_c <- rulesAsMatrix(x, "confidence")
+  m_s <- rules2matrix(x, "support")
+  m_c <- rules2matrix(x, "confidence")
   
   reorderTypes <- c("none", "measure", "support/confidence", "similarity")
   reorderType <- pmatch(reorder , reorderTypes, nomatch = 0)
@@ -258,5 +214,3 @@ matrix_plotly <- function(x, measure, shading, control, ...) {
       #yaxis=list(title="RHS", showticklabels = FALSE, showgrid = FALSE)
     )
 }
-
-